+++ src/main/java/by/it_academy/jd2/MJD29522/service/api/IGenreService.java
@@ -1,6 +1,5 @@
 package by.it_academy.jd2.MJD29522.service.api;
 
-import by.it_academy.jd2.MJD29522.dto.GenreDTO;
 import by.it_academy.jd2.MJD29522.dto.GenreID;
 
 import java.util.List;
@@ -11,9 +10,9 @@ public interface IGenreService {
 
     boolean add(String newGenre);
 
-    void update(int id, String name);
+    void update(long id, String name);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
-    boolean exist (int id);
+    boolean exist (long id);
 }
